+++ objectified-ui/public/WHATS_NEW.md
@@ -9,6 +9,7 @@ We continue to improve the platform based on your feedback with improvements and
 - Bug fixes:
   - Added support for missing patternProperties in the property editor
   - Added support for discriminator mapping in the class editor
+  - Added support for prefixItems to allow for description and examples
   - Fixed dependentProperties editing in the class editor, adds a logic builder
   - Fixed unevaluatedProperties in comprehension section of schemas
   - Fixed if/then/else in comprehension section of schemas